latest

+++ frontend/src/app/pages/events/events.component.ts
@@ -13,10 +13,11 @@ import User from "../../models/users";
 export class EventsComponent implements OnInit {
 
   addEvent!: HTMLElement;
+  confirmRegistration!: HTMLElement;
   events: Event[] = [];
   moment: any = moment;
   user!: User;
-  futureEvents: Event[]=[];
+  futureEvents: Event[] = [];
 
   constructor(private eventService: EventsService,
     private el: ElementRef,
@@ -24,6 +25,7 @@ export class EventsComponent implements OnInit {
 
   ngOnInit() {
     this.addEvent = this.el.nativeElement.querySelector("#addEvent");
+    this.confirmRegistration = this.el.nativeElement.querySelector("#confirmRegistration");
 
     const userId = localStorage.getItem("userId");
     if (userId) {
@@ -37,20 +39,55 @@ export class EventsComponent implements OnInit {
     this.eventService.geEvents()
       .subscribe(events => {
         this.events = events;
-        this.events.sort((a,b) => new Date(a.eventDateTime).valueOf() - new Date(b.eventDateTime).valueOf());
-        this.futureEvents=this.events.filter(event=>moment(event.eventDateTime)>moment());
-        console.log("future events: ",this.futureEvents)
+        this.events.sort((a, b) => new Date(a.eventDateTime).valueOf() - new Date(b.eventDateTime).valueOf());
+        this.futureEvents = this.events.filter(event => moment(event.eventDateTime) > moment());
+        console.log("future events: ", this.futureEvents)
       });
   }
 
-  public deleteEvent(id: string) {
-    this.eventService.deleteEvent(id);
+  deleteEvent(id: string) {
+
+    if (confirm("Confirm delete?")) {
+      console.log("deleting event");
+      this.eventService.deleteEvent(id)
+        .subscribe((events) => {
+          console.log("events after subscribe ", events);
+          this.events = events;
+          this.events.sort((a, b) => new Date(a.eventDateTime).valueOf() - new Date(b.eventDateTime).valueOf());
+          this.futureEvents = this.events.filter(event => moment(event.eventDateTime) > moment());
+          console.log("futureevents: ", this.futureEvents);
+        });
+    }
   }
 
-  public showAddEvent() {
+  showAddEvent() {
     this.addEvent.classList.add("w3-show");
     this.addEvent.classList.remove("w3-hide");
     //set focus to description field when add event modal is shown
     this.el.nativeElement.querySelector("#eventDescription").focus();
+  }
+
+  showConfirmRegistration() {
+    this.confirmRegistration.classList.add("w3-show");
+    this.confirmRegistration.classList.remove("w3-hide");
+    //set focus to description field when add event modal is shown
+    //this.el.nativeElement.querySelector("#eventDescription").focus();
+  }
+
+  public updateEvents(events: Event[]) {
+    this.events = events;
+    this.events.sort((a, b) => new Date(a.eventDateTime).valueOf() - new Date(b.eventDateTime).valueOf());
+    this.futureEvents = this.events.filter(event => moment(event.eventDateTime) > moment());
+    console.log("futureevents: ", this.futureEvents);
+  }
+
+  updateRegistration(eventId: string) {
+    this.showConfirmRegistration();
+    /*
+    console.log("user", this.user.displayName, "registering for ", eventId);
+
+    this.eventService.updateRegistration(eventId, this.user._id, false)
+      .subscribe((res) => console.log(res));
+      */
   }
 }
